Move TransformationRule creation into TransformingDescription
